+++ internal/platform/autostart_darwin.go
@@ -9,13 +9,13 @@ import (
 )
 
 const (
-	launchAgentLabel = "io.github.wesleiaqui.EternoMail"
+	launchAgentLabel = "io.github.wesleiaqui.eternomail"
 	launchAgentPlist = `<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
 	<key>Label</key>
-	<string>io.github.wesleiaqui.EternoMail</string>
+	<string>io.github.wesleiaqui.eternomail</string>
 	<key>ProgramArguments</key>
 	<array>
 		<string>%s</string>